+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/blockentity/RecipeRegistryBlockEntity.kt
@@ -1,0 +1,16 @@
+package site.siredvin.peripheralworks.common.blockentity
+
+import net.minecraft.core.BlockPos
+import net.minecraft.core.Direction
+import net.minecraft.world.level.block.state.BlockState
+import site.siredvin.peripheralium.common.blockentities.PeripheralBlockEntity
+import site.siredvin.peripheralworks.common.setup.BlockEntityTypes
+import site.siredvin.peripheralworks.computercraft.peripherals.RealityForgerPeripheral
+import site.siredvin.peripheralworks.computercraft.peripherals.RecipeRegistryPeripheral
+
+class RecipeRegistryBlockEntity(blockPos: BlockPos, blockState: BlockState) :
+    PeripheralBlockEntity<RecipeRegistryPeripheral>(BlockEntityTypes.RECIPE_REGISTRY.get(), blockPos, blockState) {
+    override fun createPeripheral(side: Direction): RecipeRegistryPeripheral {
+        return RecipeRegistryPeripheral(this)
+    }
+}