+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/data/ModLootTableProvider.kt
@@ -30,5 +30,6 @@ object ModLootTableProvider {
         LootTableHelper.dropSelf(consumer, Blocks.REMOTE_OBSERVER)
         LootTableHelper.dropSelf(consumer, Blocks.PERIPHERAL_PROXY)
         LootTableHelper.dropSelf(consumer, Blocks.REALITY_FORGER)
+        LootTableHelper.dropSelf(consumer, Blocks.RECIPE_REGISTRY)
     }
 }