+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/subsystem/recipe/RecipeSerializableRecord.kt
@@ -1,0 +1,5 @@
+package site.siredvin.peripheralworks.subsystem.recipe
+
+interface RecipeSerializableRecord {
+    fun serializeForToolkit(): Map<String, Any>
+}